Implemented new simpleOptions functions
